Add files via upload
This script allows you to automatically process an EMG trace and extract the MEP size (peak to peak and area under the curve), the MEP onset latency using two different methods, and the EMG activity preceding the TMS pulse, which can be useful to exclude MEPs preceded by excessive muscular contraction.


The script has two separate options to run this code: if you have a digital marker on your EMG file signalling when the TMS pulse was delivered, you can use the marker to inform the data extraction. Alternatively, if you do not have markers on the actual EMG trace but have some sort of external file (e.g., an excel file) with the timing of each TMS pulse relative to the start of the EMG recording, you can upload and use that file to do the same analysis.